--- frontend/src/Components/Wallet.js
@@ -1,9 +1,0 @@
-function Wallet({ setWalletSide }) {
-    setWalletSide(true);
-
-    return (
-        <h1>Wallet page</h1>
-    );
-}
-
-export default Wallet;
